requirements.txt
# Airflow dependencies
apache-airflow==2.7.0
apache-airflow-providers-postgres==5.6.0

# Data processing
pandas==2.0.3
numpy==1.24.4

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Database
psycopg2-binary==2.9.7

# Testing
pytest==7.4.0
pytest-cov==4.1.0 